langgraph_chart_generator/requirements.txt
langchain
langchain_openai
langgraph
fastapi
uvicorn[standard]
python-dotenv
openai
e2b-code-interpreter
requests
python-multipart 
Pillow